chore: provide dts for browser bundle
